Desktop folders/django-project-master/requirements.txt
asgiref
certifi
chardet
dj-database-url
Django
django-crispy-forms
django-heroku
django-hitcount
gunicorn
idna
Pillow
psycopg2
pytz
requests
sqlparse
urllib3
whitenoise
